requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-multipart==0.0.12
Pillow==11.0.0
